requirements.txt
# Core ML packages
torch>=2.0.0
transformers>=4.30.0

# Data processing
numpy>=1.21.0
pandas>=1.5.0

# Dashboard and visualization
streamlit>=1.25.0
plotly>=5.15.0
psutil>=5.9.0
GPUtil>=1.4.0

# Retry mechanisms
tenacity>=8.2.0

# Configuration
pyyaml>=6.0 
